Split RzBin PE parser into smaller pieces
