+++ librz/bin/format/pe/pe64_debug.c
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: 2022 Peiwei Hu <jlu.hpw@foxmail.com>
+// SPDX-License-Identifier: LGPL-3.0-only
+
+#define RZ_BIN_PE64 1
+#include "pe_debug.c"
