+++ librz/bin/meson.build
@@ -205,6 +205,28 @@ rz_bin_sources = [
   'format/pe/dotnet.c',
   'format/pe/pe.c',
   'format/pe/pe64.c',
+  'format/pe/pe_info.c',
+  'format/pe/pe64_info.c',
+  'format/pe/pe_overlay.c',
+  'format/pe/pe64_overlay.c',
+  'format/pe/pe_hdr.c',
+  'format/pe/pe64_hdr.c',
+  'format/pe/pe_imports.c',
+  'format/pe/pe64_imports.c',
+  'format/pe/pe_exports.c',
+  'format/pe/pe64_exports.c',
+  'format/pe/pe_rsrc.c',
+  'format/pe/pe64_rsrc.c',
+  'format/pe/pe_tls.c',
+  'format/pe/pe64_tls.c',
+  'format/pe/pe_security.c',
+  'format/pe/pe64_security.c',
+  'format/pe/pe_section.c',
+  'format/pe/pe64_section.c',
+  'format/pe/pe_clr.c',
+  'format/pe/pe64_clr.c',
+  'format/pe/pe_debug.c',
+  'format/pe/pe64_debug.c',
   'format/pe/pemixed.c',
   'format/pyc/marshal.c',
   'format/pyc/pyc.c',
